ref: collection editor ui
